fit long datetime

+++ models/token_info.go
@@ -11,7 +11,7 @@ package models
 
 type TokenInfo struct {
 	TokenId      string `json:"token_id,omitempty"`
-	CreationTime int32  `json:"creation_time,omitempty"`
-	ExpiryTime   int32  `json:"expiry_time,omitempty"`
+	CreationTime int64  `json:"creation_time,omitempty"`
+	ExpiryTime   int64  `json:"expiry_time,omitempty"`
 	Comment      string `json:"comment,omitempty"`
 }
